+++ stories/components/data-table/advanced/data-table.advanced.stories.tsx
@@ -2,9 +2,11 @@ import { DataTable } from '@semcore/data-table';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import AccordionInMergedRowsExample from './examples/accordion_in_merged_rows';
+import FakeMultiLineHeaderExample from './examples/fake-multi-level-header';
 import FixedColumnsWithDiffWidthsExample from './examples/fixed_columns_with_diff_widths';
 import LinkInTableExample from './examples/link_in_table';
 import OverflowInCellsExample from './examples/overflow_in_cells';
+import RenderCellRawValueExample from './examples/render_cell_raw_data';
 import RowCellStatesExample from './examples/row_cell_states';
 import SideIndentsExample from './examples/side-indents';
 
@@ -38,4 +40,12 @@ export const SideIndents: Story = {
 
 export const OverflowInCells: Story = {
   render: OverflowInCellsExample,
+};
+
+export const FakeMultiLineHeader: Story = {
+  render: FakeMultiLineHeaderExample,
+};
+
+export const RenderCellRawValue: Story = {
+  render: RenderCellRawValueExample,
 };